Somewhat Implemented. No Concrete Interceptor and not tested
The core functionality for the Interceptor stuff is there as far as I understand it. Right now only for access requests for the Database.
Will be adding the concrete interceptor to the SQLDatabase object and then the event response is any time a statement is called.
It will then write a message to the log file through the current logger.
Will be adding in the way to log the responses from the database as well.